projectfiles/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["ANRC2020 <69983424+ANRC2020@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
fastapi = {extras = ["standard"], version = "^0.112.2"}
joblib = "1.4.2"
scikit-learn = "1.5.2"
fastapi-cache2 = {version = "0.1.9", extras = ["redis"]}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.6.2"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-vv"
